bg(morgan issue): fix morgan issue
