+++ src/extends/tests/core.test.js
@@ -1,5 +1,5 @@
 import request from 'supertest';
-import app from '../../app/app';
+import app from '../../app';
 
 import { urlPrefix } from '../mocks/variables';
